Update and cleanup available RStudio containers
Attempts to fix some of the RStudio container bugs. There is still known
bugs it the R-4.3.0 version of RStudio but it now starts and correctly
finds available GPU's.

Two of the containers with R version `R-4.2.1` have been postfixed with
`.broken` as the containers start but have the wrong R version
installed. This is due to a mismatch between the containers base image
Ubuntu 20.04 and the latest available R version in the 20.04
repositories.