Add miniforge containers
Miniforge3 holds the minimal installers for Conda and Mamba specific to
conda-forge, with the following features pre-configured:

* Packages in the base environment are obtained from the conda-forge channel.
* The conda-forge channel is set as the default (and only) channel.

The container is based on a minimal Ubuntu installation that includes
conda-forges miniforge3 installer pre-installed at /opt/conda. The tags of
the images are in line with the official miniforge release tags.

The code used for building the image can be found at
https://github.com/conda-forge/miniforge-images